Adding Statistics class

+++ src/cache/memcache_element.h
@@ -6,7 +6,7 @@
 struct MemcacheElement {
     uint8_t *block;
     uint16_t flags;
-    int exptime;
+    unsigned long long int exptime;
     int bytes;
     uint64_t cas_unique;
 };
